Fixed printSolution function

+++ dijkstra.h
@@ -9,7 +9,7 @@ Labb 1
 #define DIJKSTRA_H
 #include <vector>
 //Returns the index of the node that has the least weight that is reachable and not already visited
-int minDistance(std::vector<int> dist, std::vector<bool> sptSet, int size);
+int minDistance(std::vector<double> dist, std::vector<bool> sptSet, int size);
 //Prints all paths and their corresponding total weight from a given start node
 void dijkstra(std::vector<std::vector<double> > graph, int src, int size);
 #endif
